Interface refactoring
