+++ src/main/java/czertainly/common/credential/provider/api/AttributesControllerImpl.java
@@ -1,19 +1,17 @@
 package czertainly.common.credential.provider.api;
 
-import java.util.List;
-
+import com.czertainly.api.exception.ValidationException;
+import com.czertainly.api.interfaces.connector.AttributesController;
+import com.czertainly.api.model.common.AttributeDefinition;
+import com.czertainly.api.model.common.RequestAttributeDto;
 import czertainly.common.credential.provider.service.AttributeService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.czertainly.api.exception.ValidationException;
-import com.czertainly.api.interfaces.AttributesController;
-import com.czertainly.api.model.AttributeDefinition;
+import java.util.List;
 
 @RestController
 @RequestMapping("/v1/credentialProvider/{kind}/attributes")
@@ -28,7 +26,7 @@ public class AttributesControllerImpl implements AttributesController {
     }
 
     @Override
-    public boolean validateAttributes(@PathVariable String kind, @RequestBody List<AttributeDefinition> attributes) throws ValidationException {
+    public boolean validateAttributes(@PathVariable String kind, @RequestBody List<RequestAttributeDto> attributes) throws ValidationException {
         return attributeService.validateAttributes(kind, attributes);
     }
 }
