+++ src/main/java/czertainly/common/credential/provider/service/AttributeService.java
@@ -1,11 +1,12 @@
 package czertainly.common.credential.provider.service;
 
-import java.util.List;
+import com.czertainly.api.model.common.AttributeDefinition;
+import com.czertainly.api.model.common.RequestAttributeDto;
 
-import com.czertainly.api.model.AttributeDefinition;
+import java.util.List;
 
 public interface AttributeService {
 	List<AttributeDefinition> getAttributes(String kind);
 	
-	boolean validateAttributes(String kind, List<AttributeDefinition> attributes);
+	boolean validateAttributes(String kind, List<RequestAttributeDto> attributes);
 }
